+++ app/controllers/builder/questions_controller.rb
@@ -1,12 +1,11 @@
+# frozen_string_literal: true
+
+# Allows survey editors to add and modify survey questions.
 class Builder::QuestionsController < Builder::BuilderController
   before_action :authenticate_user!
   before_action :redirect_non_builders
-
-  before_action :set_editable_survey
-  before_action :redirect_without_survey
-
-  before_action :set_editable_question,     only: [:show, :edit, :update, :destroy]
-  before_action :redirect_without_question, only: [:show, :edit, :update, :destroy]
+  before_action :find_editable_survey_or_redirect
+  before_action :find_editable_question_or_redirect, only: [:show, :edit, :update, :destroy]
 
   def index
     redirect_to builder_survey_path(@survey)
@@ -18,16 +17,11 @@ class Builder::QuestionsController < Builder::BuilderController
 
   def create
     @question = @survey.questions.where(user_id: current_user.id).new(question_params)
-
-    respond_to do |format|
-      if @question.save
-        @survey.questions << @question
-        format.html { redirect_to builder_survey_question_path(@survey, @question), notice: 'Question was successfully created.' }
-        format.json { render :show, status: :created, location: @question }
-      else
-        format.html { render :new }
-        format.json { render json: @question.errors, status: :unprocessable_entity }
-      end
+    if @question.save
+      @survey.questions << @question
+      redirect_to builder_survey_question_path(@survey, @question), notice: 'Question was successfully created.'
+    else
+      render :new
     end
   end
 
@@ -38,33 +32,34 @@ class Builder::QuestionsController < Builder::BuilderController
   end
 
   def update
-    respond_to do |format|
-      if @question.update(question_params)
-        format.html { redirect_to builder_survey_question_path(@survey, @question), notice: 'Question was successfully updated.' }
-        format.json { render :show, status: :ok, location: @question }
-      else
-        format.html { render :edit }
-        format.json { render json: @question.errors, status: :unprocessable_entity }
-      end
+    if @question.update(question_params)
+      redirect_to builder_survey_question_path(@survey, @question), notice: 'Question was successfully updated.'
+    else
+      render :edit
     end
   end
 
   def destroy
     @question.destroy
-    respond_to do |format|
-      format.html { redirect_to builder_survey_path(@survey), notice: 'Question was successfully destroyed.' }
-      format.json { head :no_content }
+    redirect_to builder_survey_path(@survey), notice: 'Question was successfully destroyed.'
+  end
+
+  # POST /questions/reorder.js
+  def reorder
+    params[:question_ids].each_with_index do |question_id, index|
+      sqo = @survey.survey_question_orders.find_by question_id: question_id
+      sqo.update question_number: index if sqo
     end
+    head :ok
   end
 
   private
 
-    def set_editable_question
-      super(:id)
-    end
+  def find_editable_question_or_redirect
+    super(:id)
+  end
 
-    def question_params
-      params.require(:question).permit(:text_en, :slug, :archived)
-    end
-
+  def question_params
+    params.require(:question).permit(:text_en, :slug, :archived)
+  end
 end