+++ test/controllers/builder/questions_controller_test.rb
@@ -8,21 +8,21 @@ class Builder::QuestionsControllerTest < ActionController::TestCase
     @question = questions(:web_question)
   end
 
-  test "should get questions as builder" do
+  test 'should get questions as builder' do
     login(@builder)
     get :index, survey_id: @survey
     assert_not_nil assigns(:survey)
     assert_redirected_to builder_survey_path(assigns(:survey))
   end
 
-  test "should not get questions as regular user" do
+  test 'should not get questions as regular user' do
     login(@regular_user)
     get :index, survey_id: @survey
     assert_nil assigns(:survey)
     assert_redirected_to root_path
   end
 
-  test "should get new question as builder" do
+  test 'should get new question as builder' do
     login(@builder)
     get :new, survey_id: @survey
     assert_not_nil assigns(:survey)
@@ -30,7 +30,7 @@ class Builder::QuestionsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not get new question as regular user" do
+  test 'should not get new question as regular user' do
     login(@regular_user)
     get :new, survey_id: @survey
     assert_nil assigns(:survey)
@@ -38,7 +38,7 @@ class Builder::QuestionsControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should create question as builder" do
+  test 'should create question as builder' do
     login(@builder)
     assert_difference('Question.count') do
       post :create, survey_id: @survey, question: { text_en: 'My New Question', slug: 'my-new-question' }
@@ -50,7 +50,7 @@ class Builder::QuestionsControllerTest < ActionController::TestCase
     assert_redirected_to builder_survey_question_path(assigns(:survey), assigns(:question))
   end
 
-  test "should not create question without text" do
+  test 'should not create question without text' do
     login(@builder)
     assert_difference('Question.count', 0) do
       post :create, survey_id: @survey, question: { text_en: '', slug: 'my-new-question' }
@@ -63,7 +63,7 @@ class Builder::QuestionsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not create question as regular user" do
+  test 'should not create question as regular user' do
     login(@regular_user)
     assert_difference('Question.count', 0) do
       post :create, survey_id: @survey, question: { text_en: 'My New Question', slug: 'my-new-question' }
@@ -73,7 +73,7 @@ class Builder::QuestionsControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should show question as builder" do
+  test 'should show question as builder' do
     login(@builder)
     get :show, survey_id: @survey, id: @question
     assert_not_nil assigns(:survey)
@@ -81,7 +81,7 @@ class Builder::QuestionsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not show question as regular user" do
+  test 'should not show question as regular user' do
     login(@regular_user)
     get :show, survey_id: @survey, id: @question
     assert_nil assigns(:survey)
@@ -89,7 +89,7 @@ class Builder::QuestionsControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should get edit question as builder" do
+  test 'should get edit question as builder' do
     login(@builder)
     get :edit, survey_id: @survey, id: @question
     assert_not_nil assigns(:survey)
@@ -97,7 +97,7 @@ class Builder::QuestionsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not get edit question as regular user" do
+  test 'should not get edit question as regular user' do
     login(@regular_user)
     get :edit, survey_id: @survey, id: @question
     assert_nil assigns(:survey)
@@ -105,7 +105,7 @@ class Builder::QuestionsControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should update question as builder" do
+  test 'should update question as builder' do
     login(@builder)
     patch :update, survey_id: @survey, id: @question, question: { text_en: 'Updated Question', slug: 'updated-question' }
     assert_not_nil assigns(:survey)
@@ -115,7 +115,7 @@ class Builder::QuestionsControllerTest < ActionController::TestCase
     assert_redirected_to builder_survey_question_path(assigns(:survey), assigns(:question))
   end
 
-  test "should not update question without name" do
+  test 'should not update question without name' do
     login(@builder)
     patch :update, survey_id: @survey, id: @question, question: { text_en: '', slug: 'updated-question' }
     assert_not_nil assigns(:survey)
@@ -126,7 +126,7 @@ class Builder::QuestionsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not update question as regular user" do
+  test 'should not update question as regular user' do
     login(@regular_user)
     patch :update, survey_id: @survey, id: @question, question: { text_en: 'Updated Question', slug: 'updated-question' }
     assert_nil assigns(:survey)
@@ -134,7 +134,7 @@ class Builder::QuestionsControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
-  test "should destroy question as builder" do
+  test 'should destroy question as builder' do
     login(@builder)
     assert_difference('Question.current.count', -1) do
       delete :destroy, survey_id: @survey, id: @question
@@ -144,7 +144,7 @@ class Builder::QuestionsControllerTest < ActionController::TestCase
     assert_redirected_to builder_survey_path(assigns(:survey))
   end
 
-  test "should not destroy question as regular user" do
+  test 'should not destroy question as regular user' do
     login(@regular_user)
     assert_difference('Question.current.count', 0) do
       delete :destroy, survey_id: @survey, id: @question
@@ -154,4 +154,9 @@ class Builder::QuestionsControllerTest < ActionController::TestCase
     assert_redirected_to root_path
   end
 
+  test 'should save question order' do
+    login(@builder)
+    post :reorder, survey_id: @survey, question_ids: [ActiveRecord::FixtureSet.identify(:web_question)], format: 'js'
+    assert_response :success
+  end
 end